+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog
 ##### master
+* carbonserver: introduce new empty-result option to allow for empty results with carbonserver #TBD
 
 ##### version 0.16.2
 * Another attempt to fix issues with release upload #449